requirements.txt
Fabric==1.10.1
PyYAML==3.11
boto==2.36.0
mock==1.0.1
netaddr==0.7.18
testfixtures==4.1.2
paramiko
troposphere>=1.0.0
